+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Bug fixes:
 - Fix `GeoDataFrame.merge()` incorrectly returning a `DataFrame` instead of a 
   `GeoDataFrame` when the `suffixes` argument is applied to the active
   geometry column (#2933).
+- Fix bug in `pandas.concat` CRS consistency checking where CRS differing by WKT 
+  whitespace only were treated as incompatible (#3023)
 
 
 ## Version 0.14 (Sep 15, 2023)