+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Development version
+
+New features and improvements:
+
+
+Bug fixes:
+
+
 ## Version 0.13 (May 6, 2023)
 
 New methods: