+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 New methods:
 
-- Added ``offset_curve`` method from shapely to GeoSeries/GeoDataframe.
+- Added ``concave_hull`` method from shapely to GeoSeries/GeoDataframe. (#2903)
+- Added ``offset_curve`` method from shapely to GeoSeries/GeoDataframe. (#2902)
 
 New features and improvements: